AutoFinGPT/requirements.txt
requests
beautifulsoup4
pandas
sentence-transformers
faiss-cpu
llama-index
transformers
scikit-learn
gym
stable-baselines3
matplotlib
shimmy
accelerate